agents/warranty-docs/tools/preferences.py
"""User preference management tool for warranty-docs agent.

Allows users to save preferences like notification settings, contact methods, etc.
"""

from typing import Any

from strands.tools import tool


@tool
def save_user_preference(
    user_id: str, preference_key: str, preference_value: Any
) -> dict[str, Any]:
    """
    Save user preference setting.

    Args:
        user_id: Unique user identifier (from Cognito or session)
        preference_key: Preference category (e.g., 'notification_email', 'contact_method', 'language')
        preference_value: Preference value to store

    Returns:
        Confirmation of saved preference

    Example:
        >>> save_user_preference('user-123', 'notification_email', 'user@example.com')
        {
            'user_id': 'user-123',
            'preference_key': 'notification_email',
            'preference_value': 'user@example.com',
            'status': 'saved',
            'message': 'Preference saved successfully'
        }
    """
    # In production, this would:
    # 1. Store in DynamoDB or RDS
    # 2. Validate user_id against Cognito
    # 3. Encrypt sensitive preferences
    # 4. Trigger notifications on preference changes

    # Validate inputs
    if not user_id or not preference_key:
        return {
            "status": "error",
            "message": "user_id and preference_key are required",
        }

    # Mock storage (in-memory for template demo)
    # Production would persist to database
    return {
        "user_id": user_id,
        "preference_key": preference_key,
        "preference_value": preference_value,
        "status": "saved",
        "message": "Preference saved successfully. Note: This is a mock implementation and will not persist across sessions.",
        "timestamp": "2025-10-29T17:30:00Z",  # In production: use datetime.utcnow()
    }
